Unify inventory to simplify vagrant commands
To execute playbooks against a vagrant machine, we have to type:

  ansible-playbook site.yml -i inventory/dev -l vagrant

With this change, we just need:

  ansible-playbook site.yml -l vagrant

And that's the same as with all others. I think the chances that we want
to provision all OFN servers at the same time are quite low. We could
introduce another group for that if we needed at.