Set 'X-Real-IP' header when reverse-proxying from nginx to Rails app.
Recommended setting; enables us to have an accurate and straightforward way of checking the client IP addresses (in production) in Rails, with: `request.headers['X-Real-IP']`. Currently we get really confusing results in Bugsnag for example, where the IP always appears to be the same, because it's the IP of the server not the client.