Do not become_user: postgres when including role
This fixes the error

```
ERROR! 'become_user' is not a valid attribute for a IncludeRole
The error appears to be in '/home/pau/dev/ofn-install/playbooks/db_integrations.yml': line 36, column 7, but may
be elsewhere in the file depending on the exact syntax problem.
```

This is consistent with the `include_role` calls we have in
`roles/db_server/` and works I managed to run this playbook against AU
production.