Reload unicorn on deploy instead of restart
This change means that if a unicorn worker is part-way through serving a request at the point of deployment, it will be restarted as soon as it's finished instead of killing the request.

Also adds a quick check for cases where unicorn is not running yet (on first deploy for example) which would otherwise throw an error (you can't reload a service that hasn't been started yet).