Fix attachment_path on fr-prod
This value is the correct attachment_path for fr-prod, but was not added in some of the recent changes. When we deployed this week it was reset to the default value, and all images were broken. It's been this way for a long time and changing it would require restructuring the paths for all images inside the S3 bucket...