Improve nginx stub_status config
This status page is only accessed internally for stats collection. It shouldn't be accessible externally. The original config here meant that nginx was binding on 0.0.0.0 port 81, which is not necessary. External access to the port was producing a 403/Forbidden response (which is not too bad), but it shouldn't leave that port open at all.