Remove unused assets path creation
We used to store assets in a shared path but now we keep the compiled
assets in each build's own directory.

This pull request doesn't remove the old directory. It will just not be
created on new servers. We could clean this up but there's probably no
harm in keeping the directory until we replace the server. In the mean
time, it serves as a backup.