Don't use Sidekiq in Travis
The Travis build environment doesn't like running Sidekiq via a service. We don't really test sidekiq or jobs processing here anyway, so it's not a huge loss.

This is the third weird issue that's come up relating to the Travis environment recently, it would be good to try switching to Github CI for running the ofn-install build at some point...