Use Python 3 in Ubuntu 20 and onwards
Starting in version 20 Python 2 packages, specially python-psycopg2
required by geerlingguy.postgresql, are no longer part of the distro and
we have to work with their versions for Python 3.

My first attempt to get this done was by setting the
`ansible_python_interpreter` to something like:

```
ansible_python_interpreter: "/usr/bin/python{% if ansible_distribution_major_version < '20' %}2.7{% else %}3{% endif%}"
```

But it dind't work  because at the time `ansible_python_interpreter`
needs to be evaluated `ansible_distribution_major_version` still is
undefined.

This is the reason I'm going for this particular more creative solution.

In the future, when the project maintainers decide to end support for
Ubuntu 16, this solution should not be necessary and we can just work
with Python 3 in Ubuntu 18 and 20.