Update postgres role and config for db integration
The db_integrations playbook was broken. I guess that the internal
workings of the postgresql role changed and it didn't expect a subtask
to be included on its own. I'm now including the whole role as in the
provisioning playbook to make it work again and reduce maintenance. It
will be a little bit slower though.