Create stats extension on all servers
It's only used for Datadog but the Rails application is tracking
extensions in its schema.rb file. Rails tries to create the extension
itself if it doesn't exist and the db user is not allowed to create
exensions at the moment. This seemed to be the easiest way but we could
also give the db user the CREATE privilege instead.