Allow Metabase to connect to managed databases
Our Metabase server has been working with replicated databases. But our
replication setup breaks regularly and we lost the knowledge how it
works.

To simplify the setup, we are going to connect Metabase directly to the
live database (with a statement timeout).

This pull request allows connecting to postgresql. We will still need to
add the user with a secret password managed in ofn-secrets.