Apply Postgres tuning to all servers
The default configuration for postgres isn't great for production. We previously applied this to larger instances on a case by case basis but it's been removed on some servers during upgrades. I've scaled it down a little bit to be more conservative and added it to the provision playbook.

The technical part here is that if Postges is performing a lot of joins or complex queries and it can't hold what it needs in buffers, it actually stops and writes things to *disk* half way through a query and has to read them from disk again before it can return the results. Potentially even multiple times per query! This can massively increase the time it takes to complete a request.

As a test I tried running some reports with a large date range as superadmin with production data to simulate worst case scenarios and got a response time up to 8 seconds. After applying these configs locally the same request dropped down to around 1 second. Basically any querying where the complexity is on the high end of the scale (like with permissions) it should make a real difference.