Extract tasks for preparing database on first deployment
This subset of tasks runs only on the first deployment, and runs before the precompile step (because precompiling will fail unless this has been done at least once). On subsequent deployments, we skip this setup and precompiling is done before the migration step to minimise the time where the database and code are out of sync during the deployment process.