Replace http with https and fake cert in local VMs
The config is smaller, easier to maintain and covers all local
containers. It also makes sure that we use the real production config
for Nginx and that the local configs don't get out of date.

It does mean that you need to accept an insecure certificate when you
access the website locally but then the connection is encrypted. This
also enables you to test browser behaviour around redirects to https and
security settings only available with https.