Add swapfile to fr_staging
Last month the memory usage hit 100% and interrupted testing.
It  hasn't happened again, but I think it's worth having as a backup. Then the server will just slow down instead of crashing, which should still be enough of a canary to warn that something's going wrong. There's plenty of free disk space.