Deploy only for self
This script template is generated and saved on each server, with the name of itself.
If the playbook was being executed on multiple servers at once, it was saving the name of all servers here, which is not right! Now it will only save the name of the current server.

It now uses the domain name instead of group name (eg staging.openfoodnetwork.org.uk instead of uk_staging).