Update node-build branch from master to main
The HEAD branch was renamed. Even after updating the name in Ansible, it
still failed on not wanting to override a tag. The workaround is to
delete the plugin remotely:

```
rm ~/.nodenv/plugins/node-build/ -rf
```

I don't think it's worth integrating into Ansible. It would be better to
switch to a bash script installing node.