Update path for certbot renewal hooks
It looks like the postgres update script wasn't run on au_prod. And
newer versions of certbot have more specific paths for renewal hooks.

I didn't bother removing the old file. It would be temporary code and
the file is not doing any harm. In fact, if one of the servers still
uses an old version of certbot, deleting the old file may break database
integrations.